fix up color selection
Add a checkbox to each color saying whether that color should be set
or not. If not, I should also disable the corresponding color picker,
but I haven't done that yet. The color-table code has been cleaned up
considerably with a Javascript function and a Jinja2 macro. The color
updating respects the hierarchy, eg True is a python-true, then a
python-constant, then a python-keyword, now that colors can be not
set.

name-constant was renamed to constant throughout.